extract replicate state
