+++ src/moonlink_service/src/test.rs
@@ -172,7 +172,7 @@ async fn test_moonlink_standalone() {
         &mut moonlink_stream,
         DATABASE.to_string(),
         TABLE.to_string(),
-        0,
+        /*lsn=*/ 1,
     )
     .await
     .unwrap();
